requirements.txt
brian2==2.9.0
brian2tools
cleosim
matplotlib==3.10.3
numpy==2.3.1
pandas
PyWavelets
scipy==1.16.0
seaborn==0.13.2
spectrum
